fix: Prevent frequent terms scoring negatively
The IDF considered the total occurrences of a token where it should have
considered the total number of documents containing the token. In
instances where the occurrences exceeded the number of documents, the
IDF (and therefore the score) would be negative. See this bug report for
more information: https://github.com/Michael-JB/bm25/pull/20

Co-authored-by: Tim Lee <hwiorn@gmail.com>